+++ src/mir/mod.rs
@@ -161,7 +161,7 @@ pub enum Op {
 
     Branch {
         cond: Symbol,
-        clauses: Vec<(u64, Symbol, bool)>,
+        clauses: Vec<(u32, Symbol, bool)>,
         default: Option<(Symbol, bool)>,
     },
     Jump {
